requirements.txt
pydicom
matplotlib
opencv-python
flask
pillow
numpy
tk